Rename infinity-list to infinite-list
