+++ packages/review/src/components/reviews.tsx
@@ -12,7 +12,7 @@ import {
   PopularReviewsInfinite,
   LatestReviewsInfinite,
   RatingReviewsInfinite,
-} from './infinity-list'
+} from './infinite-list'
 import {
   SortingOptionsProvider,
   useReviewSortingOptions,
@@ -21,7 +21,7 @@ import type { SortingOption, SortingType } from './sorting-context'
 import { FilterProvider, useReviewFilters } from './filter-context'
 import { SortingOptions } from './sorting-options'
 import { Filters } from './filter'
-import type { InfinityReviewValue } from './infinity-list'
+import type { InfinityReviewValue } from './infinite-list'
 
 const REVIEWS_SECTION_ID = 'reviews'
